Add search field and retrieve GitHub profile on click/enter
Add an `IconButton` and handle clicks on that button (or 'enter'
key-presses) by calling `searchProfiles`. This callback then
makes an API request to the GitHub API to retrieve the profile
that matches the specified username.

Also, handle loading and error states by setting these appropriately
and displaying their representation on the UI if they are set.